Update Mass Viewpoint tests
Migrate model from Capella 1.4.0 to 5.0.0.
Create proper plugin project.
