+++ massviewpoint/tests/org.polarsys.capella.vp.mass.ju/META-INF/MANIFEST.MF
@@ -1,14 +1,14 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: Ju
+Bundle-Name: Obeo
 Bundle-SymbolicName: org.polarsys.capella.vp.mass.ju
 Bundle-Version: 5.0.0.qualifier
-Bundle-ClassPath: ju.jar,
- capella-studio-1.4/plugins/org.eclipse.sirius_6.3.0.201910180815.jar
-Export-Package: org.polarsys.capella.vp.mass.ju
+Automatic-Module-Name: org.polarsys.capella.vp.mass.ju
+Bundle-RequiredExecutionEnvironment: JavaSE-14
 Require-Bundle: org.polarsys.capella.core.libraries,
- org.polarsys.capella.test.framework,
  org.polarsys.capella.core.data.gen,
- org.polarsys.capella.vp.mass
-Import-Package: org.polarsys.capella.vp.mass.mass.impl,
+ org.polarsys.capella.vp.mass,
+ org.polarsys.capella.test.framework
+Import-Package: junit.framework;version="4.13.0",
+ org.polarsys.capella.vp.mass.mass.impl,
  org.polarsys.capella.vp.mass.services
