+++ massviewpoint/tests/org.polarsys.capella.vp.mass.ju/model/MassProject/.user.af
@@ -1,0 +1,2 @@
+<?xml version="1.0" encoding="ASCII"?>
+<vp:Workspace xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:vp="http://www.polarsys.org/kitalpha/ad/viewpoint/1.0.0"/>
