+++ massviewpoint/tests/org.polarsys.capella.vp.mass.ju/model/MassProject/MassProject.afm
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<metadata:Metadata xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:metadata="http://www.polarsys.org/kitalpha/ad/metadata/1.0.0" id="_isy7QF1jEeqBCMjfTVrmpQ">
-  <viewpointReferences id="_itjJMF1jEeqBCMjfTVrmpQ" vpId="org.polarsys.capella.core.viewpoint" version="1.4.0"/>
-  <viewpointReferences id="_o0_BgF1jEeqBCMjfTVrmpQ" vpId="org.polarsys.capella.vp.mass" version="1.4.0.qualifier"/>
+<metadata:Metadata xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:metadata="http://www.polarsys.org/kitalpha/ad/metadata/1.0.0" id="_5Bl08G0QEeugr_Q04ehMGw">
+  <viewpointReferences id="_5B6lEG0QEeugr_Q04ehMGw" vpId="org.polarsys.capella.core.viewpoint" version="5.0.0"/>
+  <viewpointReferences id="_Fn_UwG0REeugr_Q04ehMGw" vpId="org.polarsys.capella.vp.mass" version="5.0.0.qualifier"/>
 </metadata:Metadata>
