+++ massviewpoint/tests/org.polarsys.capella.vp.mass.ju/src/org/polarsys/capella/vp/mass/ju/testCases/AddMassNodeTest.java
@@ -8,35 +8,31 @@
  *   Contributors:
  *      Obeo - initial API and implementation
  ******************************************************************************/
-package org.polarsys.capella.vp.mass.ju.testCase;
+package org.polarsys.capella.vp.mass.ju.testCases;
 
 import org.polarsys.capella.core.data.cs.impl.PartImpl;
 import org.polarsys.capella.core.data.pa.PhysicalComponent;
+import org.polarsys.capella.core.data.pa.impl.PhysicalComponentImpl;
 import org.polarsys.capella.vp.mass.mass.impl.PartMassImpl;
 
 /**
- * This test case checks if when editing the mass of a behavior node the mass of
- * its parents is re-calculated
- * Used to check if the listener responds to the notification SET
+ * This test case checks if after adding a mass to an already existing physical
+ * node, the mass of its parent is re-calculated
+ * Used to check if the listener responds to the notification ADD
  */
-public class EditMassBehaviorTest extends MassTest {
+public class AddMassNodeTest extends MassTest {
 	PartMassImpl pc1PartMass;
-	PartMassImpl pc121PartMass;
-
+	
 	@Override
 	public void test() throws Exception {
+		
 		PhysicalComponent pc1 = physicalSystem.getOwnedPhysicalComponents().get(0);
 		pc1PartMass = (PartMassImpl) ((PartImpl) pc1.getAbstractTypedElements().get(0)).getOwnedExtensions().get(0);
-		PhysicalComponent pc12 = pc1.getDeployedPhysicalComponents().get(0);
-		PhysicalComponent pc121 = pc12.getDeployedPhysicalComponents().get(0);
-		pc121PartMass = (PartMassImpl) ((PartImpl) pc121.getAbstractTypedElements().get(0)).getOwnedExtensions().get(0);
+		PhysicalComponentImpl pc1_3 = (PhysicalComponentImpl) pc1.getOwnedPhysicalComponents().get(1);
 
-		// change the mass of pc1.2.1 (a behavior) and check if the mass of pc1 is
-		// re-calculated
-		editMassPhysicalComponent(pc121, 10);
-		assertEquals("The mass of PC1.2.1 was not changed after its modification", pc121PartMass.getCurrentMass(), 10);
-		assertEquals("The mass of PC1 was not changed after the modification of pc1.2.1", pc1PartMass.getCurrentMass(),
-				90);
+		// add a mass to pc1.3 and check if the mass of pc1 is re-calculated
+		addMassToPhysicalComponent(pc1_3, 10);
+		assertEquals("The mass of PC1 was not changed after the addition of pc1.3", pc1PartMass.getCurrentMass(), 95);
 
 	}
