+++ massviewpoint/tests/org.polarsys.capella.vp.mass.ju/src/org/polarsys/capella/vp/mass/ju/testCases/DeleteBehaviorWithMassTest.java
@@ -8,7 +8,7 @@
  *   Contributors:
  *      Obeo - initial API and implementation
  ******************************************************************************/
-package org.polarsys.capella.vp.mass.ju.testCase;
+package org.polarsys.capella.vp.mass.ju.testCases;
 
 import org.polarsys.capella.common.ef.ExecutionManager;
 import org.polarsys.capella.common.ef.command.AbstractReadWriteCommand;
@@ -19,33 +19,37 @@ import org.polarsys.capella.test.framework.helpers.EObjectHelper;
 import org.polarsys.capella.vp.mass.mass.impl.PartMassImpl;
 
 /**
- * This test case checks if when deleting a partMass of a node PC the mass of
- * its parents is re-calculated.
+ * This test case checks if when deleting a behavior PC owning a partMass the
+ * mass of its parents is re-calculated.
  * Used to check if the listener responds to the notification REMOVE
  */
-public class DeletePartMassNodeTest extends MassTest {
+public class DeleteBehaviorWithMassTest extends MassTest {
 
-	private PartMassImpl pc1PartMass;
-	private PartMassImpl pc11PartMass;
+	PartMassImpl pc1PartMass;
+	PartMassImpl pc12PartMass;
+	PhysicalComponent pc12;
 
 	@Override
 	public void test() throws Exception {
+
 		PhysicalComponent pc1 = physicalSystem.getOwnedPhysicalComponents().get(0);
 		pc1PartMass = (PartMassImpl) ((PartImpl) pc1.getAbstractTypedElements().get(0)).getOwnedExtensions().get(0);
-		PhysicalComponent pc11 = pc1.getOwnedPhysicalComponents().get(0);
-		pc11PartMass = (PartMassImpl) ((PartImpl) pc11.getAbstractTypedElements().get(0)).getOwnedExtensions().get(0);
+		pc12 = pc1.getDeployedPhysicalComponents().get(0);
+		pc12PartMass = (PartMassImpl) ((PartImpl) pc12.getAbstractTypedElements().get(0)).getOwnedExtensions().get(0);
 
 		// remove pc1.2 and check if the mass of pc1 is re-calculated
-		ExecutionManager manager = TransactionHelper.getExecutionManager(pc11PartMass);
+		ExecutionManager manager = TransactionHelper.getExecutionManager(pc12);
 		manager.execute(new AbstractReadWriteCommand() {
 
 			@Override
 			public void run() {
-				EObjectHelper.removeElement(pc11PartMass);
+				EObjectHelper.removeElement(pc12);
 			}
 
 		});
-		assertEquals("The mass of PC1 was not changed after the deletion of the part mass of PC 1.1",
-				pc1PartMass.getCurrentMass(), 35);
+		
+		assertEquals("The mass of PC1 was not changed after the deletion of pc1.2", pc1PartMass.getCurrentMass(), 70);
+
 	}
+
 }
