+++ massviewpoint/tests/org.polarsys.capella.vp.mass.ju/src/org/polarsys/capella/vp/mass/ju/testCases/DeleteNodeWithMassTest.java
@@ -8,7 +8,7 @@
  *   Contributors:
  *      Obeo - initial API and implementation
  ******************************************************************************/
-package org.polarsys.capella.vp.mass.ju.testCase;
+package org.polarsys.capella.vp.mass.ju.testCases;
 
 import org.polarsys.capella.common.ef.ExecutionManager;
 import org.polarsys.capella.common.ef.command.AbstractReadWriteCommand;
@@ -19,35 +19,31 @@ import org.polarsys.capella.test.framework.helpers.EObjectHelper;
 import org.polarsys.capella.vp.mass.mass.impl.PartMassImpl;
 
 /**
- * This test case checks if when deleting the partMass of a behavior PC the mass
- * of its parents is re-calculated.
+ * This test case checks if when deleting a node the mass of its parent is
+ * re-calculated.
  * Used to check if the listener responds to the notification REMOVE
  */
-public class DeletePartMassBehaviorTest extends MassTest {
-
-	private PartMassImpl pc1PartMass;
-	private PhysicalComponent pc12;
-	private PartMassImpl pc12PartMass;
+public class DeleteNodeWithMassTest extends MassTest {
+	PartMassImpl pc1PartMass;
 
 	@Override
 	public void test() throws Exception {
 		PhysicalComponent pc1 = physicalSystem.getOwnedPhysicalComponents().get(0);
 		pc1PartMass = (PartMassImpl) ((PartImpl) pc1.getAbstractTypedElements().get(0)).getOwnedExtensions().get(0);
-		pc12 = pc1.getDeployedPhysicalComponents().get(0);
-		pc12PartMass = (PartMassImpl) ((PartImpl) pc12.getAbstractTypedElements().get(0)).getOwnedExtensions().get(0);
+		PhysicalComponent pc11 = pc1.getOwnedPhysicalComponents().get(0);
 
-		// remove pc12PartMass and check if the mass of pc1 is re-calculated
-		ExecutionManager manager = TransactionHelper.getExecutionManager(pc12PartMass);
+		// remove pc1.1 and check if the mass of pc1 is re-calculated
+		ExecutionManager manager = TransactionHelper.getExecutionManager(pc11);
 		manager.execute(new AbstractReadWriteCommand() {
 
 			@Override
 			public void run() {
-				EObjectHelper.removeElement(pc12PartMass);
+				EObjectHelper.removeElement(pc11);
 			}
 
 		});
-		assertEquals("The mass of PC1 was not changed after the deletion of the part mass of PC 1.2",
-				pc1PartMass.getCurrentMass(), 75);
+		assertEquals("The mass of PC1 was not changed after the deletion of pc1.1", pc1PartMass.getCurrentMass(), 35);
+
 	}
 
 }
