+++ massviewpoint/tests/pom.xml
@@ -5,7 +5,6 @@
   are made available under the terms of the Eclipse Public License v1.0
   which accompanies this distribution, and is available at
   http://www.eclipse.org/legal/epl-v10.html
-
   Contributors:
        Obeo - initial API and implementation
 -->
